# packages/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "OAuthAuthorizeInput": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OAuthAuthorizeInput.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": ["array", "null"]
          },
          "state": {
            "type": "string"
          }
        },
        "required": ["client_id", "redirect_uri", "scopes", "org_id"],
        "type": "object"
      },
      "OAuthAuthorizeResponse": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OAuthAuthorizeResponse.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "redirect_to": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": ["redirect_to", "code"],
        "type": "object"
      },
...
      "OAuthTokenResponse": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OAuthTokenResponse.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "format": "int64",
            "type": "integer"
          },
          "refresh_token": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        },
        "required": ["access_token", "token_type", "expires_in", "scope"],
        "type": "object"
      },
      "OauthAuthorizeInfoResponse": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OauthAuthorizeInfoResponse.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "client_name": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": ["array", "null"]
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object"
      },
      "OauthRevokeRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OauthRevokeRequest.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": ["token"],
        "type": "object"
      },
      "OauthTokenRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OauthTokenRequest.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "client_secret": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "grant_type": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": ["grant_type", "client_id", "client_secret"],
        "type": "object"
      },
...
      "OpenAPIError": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OpenAPIError.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": ["error"],
        "type": "object"
      },
...
      "User": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": ["id", "email", "created_at"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT or personal access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Nanoflare Control API",
    "version": "v1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/oauth/authorize": {
      "get": {
        "operationId": "getOAuthAuthorization",
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "schema": {
              "format": "uri",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OauthAuthorizeInfoResponse"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "getOAuthAuthorization",
        "tags": ["OAuth"]
      },
      "post": {
        "operationId": "authorizeOAuthClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthAuthorizeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAuthorizeResponse"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "authorizeOAuthClient",
        "tags": ["OAuth"]
      }
    },
    "/v1/oauth/revoke": {
      "post": {
        "operationId": "revokeOAuthToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OauthRevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "revokeOAuthToken",
        "tags": ["OAuth"]
      }
    },
    "/v1/oauth/token": {
      "post": {
        "operationId": "createOAuthToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OauthTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthTokenResponse"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "createOAuthToken",
        "tags": ["OAuth"]
      }
    },
...
  }
}